% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_params.R
\name{assign_params}
\alias{assign_params}
\title{assign_params}
\usage{
assign_params(.list, .par_names, .env)
}
\arguments{
\item{.list}{\verb{named list}}

\item{.par_names}{\code{vector (chr)}}

\item{.env}{un \link{envir} o \code{NULL}}
}
\value{
\code{NULL}
}
\description{
Assegna a nuove variabili in \code{env} gli elementi
con nome \code{.par_names} presenti nella lista \code{.list}
}
