% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env_variables.R
\name{carica_env_vars}
\alias{carica_env_vars}
\title{carica_env_vars}
\usage{
carica_env_vars(.link_postgres_locale)
}
\arguments{
\item{.link_postgres_locale}{\code{chr} (SI/NO)}
}
\value{
\code{NULL}
}
\description{
Crea le variabili d'ambiente per VM e Airflow se \code{.link_postgres_locale == "NO"}
o per PC locale se \code{.link_postgres_locale == "SI"}
}
