% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carica_parametri_input.R
\name{carica_parametri_input}
\alias{carica_parametri_input}
\title{carica_parametri_input}
\usage{
carica_parametri_input(
  .lancio_interfaccia,
  .id_elaborazione,
  .out_version,
  .dat_report,
  .codice_utente
)
}
\arguments{
\item{.lancio_interfaccia}{\code{chr} (SI/NO)}

\item{.id_elaborazione}{\code{int}}

\item{.out_version}{\code{int}}

\item{.dat_report}{\verb{date (\%Y-\%m-\%d)}}

\item{.codice_utente}{\code{chr}}
}
\value{
\code{list}
}
\description{
Inizializza i parametri di input
}
