% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect2bq.R
\name{connect2bq}
\alias{connect2bq}
\title{connect2bq}
\usage{
connect2bq(.project_id, .dataset)
}
\arguments{
\item{.project_id}{\code{chr} nome del progetto}

\item{.dataset}{\code{chr} nome del dataset}
}
\value{
Un S4 object ereditato da \link[DBI]{dbConnect}, usato per comunicare con il database
}
\description{
Crea una connessione a Big Query
}
