% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect2postgres.R
\name{connect2postgres}
\alias{connect2postgres}
\title{connect2postgres}
\usage{
connect2postgres(.host, .port, .dbname, .user, .password)
}
\arguments{
\item{.host}{\code{chr}}

\item{.port}{\code{chr}}

\item{.dbname}{\code{chr}}

\item{.user}{\code{chr}}

\item{.password}{\code{chr}}
}
\value{
Un S4 object ereditato da \link[DBI]{dbConnect}, usato per comunicare con il database
}
\description{
Crea una connessione al database \code{ddl-dcerm-db} su Postgres
}
