% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_conn.R
\name{create_conn}
\alias{create_conn}
\title{create_conn.R}
\usage{
create_conn()
}
\value{
Un S4 object ereditato da \link[DBI]{dbConnect}, usato per comunicare con il database.
}
\description{
Crea una connessione a Postgres
}
