% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_dq.R
\name{do_dq}
\alias{do_dq}
\title{do_dq}
\usage{
do_dq(
  .dq,
  .cod_processo,
  .cod_severita,
  .cod_controllo,
  .des_controllo,
  .msg,
  .con_bq,
  .con_postgres,
  .params_config
)
}
\arguments{
\item{.cod_processo}{\code{chr} codice del processo}

\item{.cod_severita}{\code{chr} codice severità del dq}

\item{.cod_controllo}{\code{chr} codice del controllo}

\item{.des_controllo}{\code{chr} descrizione del controllo}

\item{.msg}{\code{chr} messaggio associato al controllo da scrivere su Big Query}

\item{.con_bq}{\verb{S4 object} usato per comunicare con il database Big Query}

\item{.con_postgres}{\verb{S4 object} usato per comunicare con il database Postgres}

\item{.params_config}{\code{list} contiene i parametri aggiunti specifici del controllo
(tutti i parametri inclusi tra {} in msg, \code{dat_report} e \code{out_versione})}

\item{.df_errors}{può essere un \code{tibble}:
\itemize{
\item \code{output1} \code{int}
\item \code{output2} \code{int} opzionale
\item \code{output3} \code{int} opzionale
\item \code{conteggio} \code{int}
oppure una funzione che restituisce \code{df_errors} associato al data quality
}}
}
\value{
\code{chr} OK/KO
}
\description{
Esegue il data quality associato al \code{cod_controllo} del processo
\code{cod_processo}
}
