% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_query.R
\name{do_query}
\alias{do_query}
\title{do_query}
\usage{
do_query(.con, .query)
}
\arguments{
\item{.con}{\verb{S4 object} usato per comunicare con il database}

\item{.query}{\code{chr} query da eseguire su Postgres.}
}
\value{
\code{tibble}
}
\description{
Funzione che esegue una query data in input
}
