% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_scrittura_output.R
\name{do_scrittura_output}
\alias{do_scrittura_output}
\title{do_scrittura_output}
\usage{
do_scrittura_output(.df_config, .con_bigquery, .out_version, .dat_report)
}
\arguments{
\item{.df_config}{\code{tibble} le cui righe contengono le informazioni per la scrittura
di ciascun data frame:
\itemize{
\item \code{df_name} \code{chr}
\item \code{df} \code{list (tibble)}
\item \code{project_id} \code{chr}
\item \code{dataset} \code{chr}
\item \code{tabella} \code{chr}
\item \code{campi_required} \code{chr}
}}

\item{.con_bigquery}{\verb{S4 object} usato per comunicare con il database Big Query}

\item{.out_version}{\code{int}}

\item{.dat_report}{\verb{date (\%Y-\%m-\%d)}}
}
\value{
\code{vector (boolean)} indica per ciascun dataframe se la scrittura è andata a
buon fine (\code{TRUE})
}
\description{
Scrive le tabelle risultanti dall'esecuzione della business logic
su Big Query
}
