% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_dferrors.R
\name{eval_dferrors}
\alias{eval_dferrors}
\title{eval_dferrors}
\usage{
eval_dferrors(
  .df_errors,
  .cod_processo,
  .cod_severita,
  .cod_controllo,
  .des_controllo,
  .msg,
  .con_bq,
  .con_postgres,
  .params_config
)
}
\arguments{
\item{.df_errors}{\code{tibble}:
\itemize{
\item \code{output1} \code{int} o \code{int}
\item \code{output2} \code{int} o \code{int} opzionale
\item \code{output3} \code{int} o \code{int} opzionale
\item \code{conteggio} \code{int}
}}

\item{.cod_processo}{\code{chr} codice del processo}

\item{.cod_severita}{\code{chr} codice severità del dq}

\item{.cod_controllo}{\code{chr} codice del controllo}

\item{.des_controllo}{\code{chr} descrizione del controllo}

\item{.msg}{\code{chr} messaggio associato al controllo da scrivere su Big Query}

\item{.con_bq}{\verb{S4 object} usato per comunicare con il database Big Query}

\item{.con_postgres}{\verb{S4 object} usato per comunicare con il database Postgres}

\item{.params_config}{\code{list} contiene i parametri aggiunti specifici del controllo
(tutti i parametri inclusi tra {} in msg, \code{dat_report} e \code{out_version})}
}
\value{
\code{chr} OK/KO
}
\description{
Elabora il \code{df_errors} associato al \code{cod_controllo} del processo \code{cod_processo}
per la scrittura su Big Query e Postgres
}
