% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_df.R
\name{extract_df}
\alias{extract_df}
\title{extract_df}
\usage{
extract_df(.con, .table_name, .table_version, .dat_report, .order_input = NULL)
}
\arguments{
\item{.con}{un oggetto output di \link{connect2bq}, usato per comunicare con il database}

\item{.table_name}{\code{chr} nome della tabella}

\item{.table_version}{\code{int} versione della tabella}

\item{.dat_report}{\verb{date (\%Y-\%m-\%d)}}

\item{.order_input}{\code{vector} vettore di \code{chr} contenente l'ordinamento delle colonne. Se non specificato
restituisce l'ordinamento di default di Big Query}
}
\value{
\code{tibble}
}
\description{
Estrae una specifica versione (\code{.table_version}) dalla tabella (\code{.table_version})
in BigQuery, filtrata per una specifica \code{.dat_report}
}
