% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_dferrors.R
\name{mutate_dferrors}
\alias{mutate_dferrors}
\title{mutate_dferrors}
\usage{
mutate_dferrors(
  .df_errors,
  .cod_processo,
  .cod_severita,
  .cod_controllo,
  .des_controllo,
  .msg,
  .params_config
)
}
\arguments{
\item{.df_errors}{\code{tibble}:
\itemize{
\item \code{output1} \code{int}
\item \code{output2} \code{int} opzionale
\item \code{output3} \code{int} opzionale
\item \code{conteggio} \code{int}
}}

\item{.cod_processo}{\code{chr} codice del processo}

\item{.cod_severita}{\code{chr} codice severità del dq}

\item{.cod_controllo}{\code{chr} codice del controllo}

\item{.des_controllo}{\code{chr} descrizione del controllo}

\item{.msg}{\code{chr} messaggio associato al controllo da scrivere su Big Query}

\item{.params_config}{\code{list} contiene i parametri aggiunti specifici del controllo
(tutti i parametri inclusi tra {} in msg, \code{dat_report} e \code{out_versione})}
}
\value{
\code{tibble}:
\itemize{
\item \code{COD_PROCESSO} \code{chr}
\item \code{COD_SEVERITA} \code{chr}
\item \code{COD_CONTROLLO} \code{chr}
\item \code{DES_CONTROLLO} \code{chr}
\item \code{DAT_REPORT} \verb{date (\%Y-\%m-\%d)}
\item \code{ID_VERSIONE} \code{int}
\item \code{DAT_INSERIMENTO} \verb{date (\%Y-\%m-\%d)}
}
}
\description{
Modifica \code{df_errors} per la scrittura su Postgres
}
