% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_out_bl.R
\name{mutate_out_bl}
\alias{mutate_out_bl}
\title{mutate_out_bl}
\usage{
mutate_out_bl(.df_list, .df_name, .out_version, .dat_report)
}
\arguments{
\item{.df_list}{\verb{named list} contiene i data frame}

\item{.df_name}{\code{vector (chr)} nomi dei data frame}

\item{.out_version}{\code{int}}

\item{.dat_report}{\verb{date (\%Y-\%m-\%d)}}
}
\value{
\code{list} contiene i data frame
}
\description{
Aggiunge le colonne \code{DAT_REPORT} e \code{ID_VERSIONE} ai \code{df_name}
presenti in \code{df_list}
}
