% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{out2log_local}
\alias{out2log_local}
\title{out2log_local.R}
\usage{
out2log_local(...)
}
\description{
Usa \link[base]{cat} per scrivere messaggi su stdout, così da essere utile in run locali (lancio_interfaccia == NO).
}
