% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_postgres.R
\name{query_postgres}
\alias{query_postgres}
\title{query_postgres.R}
\usage{
query_postgres(query)
}
\arguments{
\item{query}{chr, query da eseguire su Postgres.}
}
\value{
Una tibble con il risultato della query.
}
\description{
Funzione che esegue una query data in input in Postgres.
}
