% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_catalog.R
\name{retrieve_data_catalog}
\alias{retrieve_data_catalog}
\title{retrieve_data_catalog}
\usage{
retrieve_data_catalog(.cod_processo)
}
\arguments{
\item{.cod_processo}{\code{chr} codice del processo per il quale sono richiesti i controlli}

\item{.catalog_table}{\code{chr} nome tabella Postgres contenente i controlli di data quality}

\item{project_id}{\code{chr} nome del progetto GCP}
}
\value{
\code{tibble}:
\itemize{
\item COD_CONTROLLO  \code{chr}
\item COD_PROCESSO \code{chr}
\item COD_TIPO_CONTROLLO \code{chr}
\item COD_FREQ_CONTROLLO \code{chr}
\item COD_SEVERITA \code{chr}
\item DES_CONTROLLO \code{chr}
\item MSG \code{chr}
\item STMT_CONTROLLO \code{chr}
\item FLG_ATTIVO \code{chr}
\item TMS_CREAZIONE ?
\item DAT_FINE_VALIDITA ?
\item COD_UTENTE_MODIFICA ?
\item COD_UTENTE_CREAZIONE \code{chr}
}
}
\description{
Esegue il download dei controlli di data quality da Postgres in base al processo indicato
}
