% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_params.R
\name{retrieve_params}
\alias{retrieve_params}
\title{retrieve_params}
\usage{
retrieve_params(.con, .id_elaborazione)
}
\arguments{
\item{.con}{.con un oggetto output di \link{connect2postgres}, usato per comunicare con il database}

\item{.id_elaborazione}{\code{int} id elaborazione riferito al singolo processo}

\item{.dataset}{\code{chr} nome del dataset}
}
\value{
\code{tibble} con il risultato della query, da interrogare con \verb{[param] = value}
}
\description{
Recupera dei parametri necessari all'esecuzione del processo
}
