% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.list.R
\name{tbl.list}
\alias{tbl.list}
\title{tbl.list}
\usage{
\method{tbl}{list}(src, df_name)
}
\arguments{
\item{src}{\code{list}}

\item{df_name}{\code{chr}}
}
\value{
\code{tibble}
}
\description{
Crea un metodo \link[dplyr]{tbl} per la classe \code{list} che
restituisce un \code{tibble} da una lista di \code{tibble}
}
