% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_interfaccia_functions.R
\name{valore_campo_string_startwith}
\alias{valore_campo_string_startwith}
\title{valore_campo_string_startwith}
\usage{
valore_campo_string_startwith(df_table, col_check, val_string)
}
\description{
tba
}
