% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df_2_BQ.R
\name{write_df_2_BQ}
\alias{write_df_2_BQ}
\title{write_df_2_BQ}
\usage{
write_df_2_BQ(project_id, dataframe, dataset, tabella, tab_schema = NULL)
}
\arguments{
\item{project_id}{chr, progetto contenente la base dati sorgente.}

\item{dataframe}{tibble, dataframe da scrivere.}

\item{dataset}{chr, dataset di destinazione.}

\item{tabella}{chr, tabella di destinazione}

\item{tab_schema}{list, lista con la stessa struttura dell'output schema_fields.
Se non specificato valore di default nullo e viene eseguito l'if.}
}
\value{
?
}
\description{
Funzione per l'inserimento di dati su BIGQUERY.
}
