% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writedf2bq.R
\name{writedf2bq}
\alias{writedf2bq}
\title{writedf2bq}
\usage{
writedf2bq(
  .project_id,
  .dataframe,
  .dataset,
  .tabella,
  .out_version,
  .dat_report,
  .tab_schema = NULL
)
}
\arguments{
\item{.project_id}{\code{chr} nome progetto}

\item{.dataframe}{\code{tibble} tibble da scrivere}

\item{.dataset}{\code{chr} dataset di destinazione}

\item{.tabella}{\code{chr} tabella di destinazione}

\item{.out_version}{\code{int} ID_VERSIONE da inserire nel dataframe}

\item{.dat_report}{\code{chr} DAT_REPORT da inserire nel dataframe}

\item{.tab_schema}{uno schema object che può essere passato come argomento di \link[bigQueryR]{bqr_upload_data}}
}
\value{
\code{bool} TRUE, invisibilmente.
}
\description{
Funzione per l'inserimento di dati su Big Query
}
