% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writedf2postgres.R
\name{writedf2postgres}
\alias{writedf2postgres}
\title{writedf2postgres}
\usage{
writedf2postgres(.nome_tabella, .dataframe)
}
\arguments{
\item{.nome_tabella}{\code{chr} nome tabella di destinazione}

\item{.dataframe}{\code{tibble} tibble da scrivere}
}
\value{
\code{bool} TRUE, invisibilmente.
}
\description{
Funzione per appendere un dataframe a una tabella del database \code{ddl-dcerm-db} su Postgres
}
