% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect2postgres.R
\name{connect2postgres}
\alias{connect2postgres}
\title{connect2postgres}
\usage{
connect2postgres()
}
\value{
Un S4 object ereditato da \link[DBI]{dbConnect}, usato per comunicare con il database
}
\description{
Crea una connessione al database \code{ddl-dcerm-db} su Postgres
}
