% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_dq.R
\name{do_dq}
\alias{do_dq}
\title{do_dq}
\usage{
do_dq(
  .catalog,
  .con,
  .con_bigquery,
  .con_postgres,
  .params_config,
  .lst_df_errors = NULL
)
}
\arguments{
\item{.catalog}{\code{tibble} contiene tutti controlli del processo:
\itemize{
\item \code{cod_controllo} \code{chr}
\item \code{cod_processo} \code{chr}
\item \code{cod_tipo_controllo} \code{int}
\item \code{cod_freq_controllo} \code{chr}
\item \code{cod_severita} \code{chr}
\item \code{des_controllo} \code{chr}
\item \code{msg} \code{chr}
\item \code{stmt_controllo} \code{chr}
\item \code{flg_attivo} \code{chr}
\item \code{tms_creazione} ?
\item \code{tms_modifica} ?
\item \code{dat_fine_validita} ?
\item \code{cod_utente_modifica} ?
\item \code{cod_utente_creazione} ?
\item \code{nome_funzione} \code{chr}
\item \code{check_interfaccia} \code{chr}
}}

\item{.con}{\verb{S4 object} usato per comunicare con il database Big Query o una lista
di \code{tibble} nella fase di esecuzione delle funzioni contenute in \code{nome_funzione}}

\item{.con_bigquery}{\verb{S4 object} usato per comunicare con il database Big Query}

\item{.con_postgres}{\verb{S4 object} usato per comunicare con il database Postgres}

\item{.params_config}{\code{list} contiene i parametri aggiuntivi necessari per eseguire
le funzioni di data quality}

\item{.lst_df_errors}{\verb{named list} lista di \code{df_errors} relativi ai data quality
eseguiti durante la business logic, in cui il nome di ciascun data frame è il
rispettivo codice controllo associato}
}
\value{
\code{tibble}:
\itemize{
\item \code{cod_processo} \code{chr}
\item \code{cod_controllo} \code{chr}
\item \code{cod_severita} \code{int}
\item \code{esito} \code{chr}
\item \code{des_controllo} \code{chr}
}
}
\description{
Esegue i data quality presenti nel catalog
}
