% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_DQ.R
\name{do_DQ}
\alias{do_DQ}
\title{do_DQ.R}
\usage{
do_DQ(check, catalog, project_id, out_version, df_errors = NULL)
}
\arguments{
\item{check}{chr, codice controllo da eseguire.}

\item{catalog}{tibble, contiene tutti controlli del processo.}

\item{project_id}{chr, nome del progetto GCP}

\item{out_version}{int, versione dell'output}

\item{df_error}{tibble con almeno 1 colonna: !!! TODO: aggiungere tipi !!!
\itemize{
\item OUTPUT1,
\item OUTPUT2 opzionale,
\item OUTPUT3 opzionale,
\item CONTEGGIO opzionale.
}}

\item{y}{A number.}
}
\value{
Un booleano. \code{TRUE} in presenza di errore nel controllo.
}
\description{
Esegue il controllo corrispondente a \code{COD_CONTROLLO}.
Ci sono due possibilità:
\itemize{
\item controllo tramite query. \code{catalog} deve avere popolato il campo \code{STMT_CONTROLLO}
con la query che verrà utilizzata per il controllo e \code{df_errors=NULL}.
\item controllo diretto se \code{df_error} è passato come argomento.
}
}
