% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_postgres.R
\name{query_postgres}
\alias{query_postgres}
\title{query_postgres}
\usage{
query_postgres(.con, .query)
}
\arguments{
\item{.con}{un oggetto output di \link{connect2postgres}, usato per comunicare con il database}

\item{query}{\code{chr} query da eseguire su Postgres.}
}
\value{
\code{tibble}
}
\description{
Funzione che esegue una query data in input in Postgres
}
