% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_catalog.R
\name{retrieve_data_catalog}
\alias{retrieve_data_catalog}
\title{retrieve_data_catalog}
\usage{
retrieve_data_catalog(.con, .cod_processo, .data_quality_postgres)
}
\arguments{
\item{.con}{un oggetto output di \link{connect2postgres}, usato per comunicare con il database}

\item{.cod_processo}{\code{chr} codice del processo per il quale sono richiesti i controlli}

\item{.data_quality_postgres}{\code{chr}}
}
\value{
\code{tibble}:
\itemize{
\item COD_CONTROLLO  \code{chr}
\item COD_PROCESSO \code{chr}
\item COD_TIPO_CONTROLLO \code{chr}
\item COD_FREQ_CONTROLLO \code{chr}
\item COD_SEVERITA \code{chr}
\item DES_CONTROLLO \code{chr}
\item MSG \code{chr}
\item STMT_CONTROLLO \code{chr}
\item FLG_ATTIVO \code{chr}
\item TMS_CREAZIONE ?
\item DAT_FINE_VALIDITA ?
\item COD_UTENTE_MODIFICA ?
\item COD_UTENTE_CREAZIONE \code{chr}
}
}
\description{
Esegue il download dei controlli di data quality da Postgres
(\code{.data_quality_postgres == "SI"})  o carica
i controlli di data quality dal file in locale (\code{.data_quality_postgres == "NO"})
in base al processo indicato
}
