% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_catalog.R
\name{retrieve_data_catalog}
\alias{retrieve_data_catalog}
\title{retrieve_data_catalog.R}
\usage{
retrieve_data_catalog(cod_processo, catalog_table)
}
\arguments{
\item{cod_processo}{chr, codice del processo per il quale sono richiesti i controlli.}

\item{catalog_table}{tibble, tabella Postgres contenente i controlli di data quality.}

\item{project_id}{chr, nome del progetto GCP.}
}
\value{
Una tibble con 13 colonne:
\itemize{
\item COD_CONTROLLO  chr,
\item COD_PROCESSO chr,
\item COD_TIPO_CONTROLLO chr,
\item COD_FREQ_CONTROLLO chr,
\item COD_SEVERITA chr,
\item DES_CONTROLLO chr,
\item MSG chr,
\item STMT_CONTROLLO chr,
\item FLG_ATTIVO chr,
\item TMS_CREAZIONE ?,
\item DAT_FINE_VALIDITA ?,
\item COD_UTENTE_MODIFICA ?,
\item COD_UTENTE_CREAZIONE chr.
}
}
\description{
Esegue il download dei controlli di data quality da Postgres in base al processo indicato.
}
