% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df_on_postgres.R
\name{write_df_on_postgres}
\alias{write_df_on_postgres}
\title{write_df_on_postgres}
\usage{
write_df_on_postgres(nome_tabella, oggetto)
}
\arguments{
\item{nome_tabella}{chr, nome tabella Postgres di destinazione.}

\item{oggetto}{tibble, oggetto da scrivere.}

\item{project_id}{chr, nome del progetto GCP.}
}
\value{
?
}
\description{
Funzione per l'inserimento di dati su Postgres.
}
